refactor: server index file path change
