+++ server/src/controllers/code-review/code-review.controller.js
@@ -1,6 +1,6 @@
-import { DateValidation, PaginationValidation } from '##/validations/index.js';
+import { DateValidation, PaginationValidation } from '##/validations/validations.js';
 import { VersionControl } from '##/use-cases/version-control/version-control.js';
-import { STATUS_CODE } from '##/constants/index.js';
+import { STATUS_CODE } from '##/constants/constants.js';
 
 export class CodeReviewController {
   static async getCodeReview(req, res) {
