+++ server/src/frameworks/express-web-server/app.js
@@ -4,10 +4,10 @@ import { fileURLToPath } from 'url';
 import helmet from 'helmet';
 
 import { ServerConfiguration } from '##/configs/server.config.js';
-import { NODE_ENVIRONMENT_MODE, STATUS_CODE } from '##/constants/index.js';
+import { NODE_ENVIRONMENT_MODE, STATUS_CODE } from '##/constants/constants.js';
 
 import { router } from './routes/routes.js';
-import { globalErrorHandlerMiddleware } from './middlewares/index.js';
+import { globalErrorHandlerMiddleware } from './middlewares/middlewares.js';
 
 const dirName = path.dirname(fileURLToPath(import.meta.url));
 const app = express();
