+++ server/src/frameworks/express-web-server/routes/code-review/code-review.route.js
@@ -1,7 +1,7 @@
 import express from 'express';
 
-import { CodeReviewController } from '##/controllers/index.js';
-import { catchAsync } from '##/utils/index.js';
+import { CodeReviewController } from '##/controllers/controllers.js';
+import { catchAsync } from '##/utils/utils.js';
 
 const router = express.Router();
