+++ server/src/frameworks/express-web-server/routes/open-api-spec/open-api-spec.route.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import swaggerUi from 'swagger-ui-express';
 
-import { OpenAPISpecController } from '##/controllers/index.js';
+import { OpenAPISpecController } from '##/controllers/controllers.js';
 
 const router = express.Router();
