+++ server/tests/integration/common-tests/date-pagination-tests.js
@@ -1,7 +1,7 @@
 import request from 'supertest';
 import { it, expect } from '@jest/globals';
 
-import { STATUS_CODE } from '##/constants/index.js';
+import { STATUS_CODE } from '##/constants/constants.js';
 import {
   PAGINATION_SIZE_MUST_BE_NUMBER,
   PAGINATION_CURSOR_SIZE_MUST_BE_NUMBER,
