+++ server/tests/integration/trunk-based-development/pull-requests/pull-requests.test.js
@@ -2,11 +2,11 @@ import request from 'supertest';
 import { jest, describe, it, expect } from '@jest/globals';
 
 import app from '##/frameworks/express-web-server/app.js';
-import { AppError } from '##/utils/index.js';
+import { AppError } from '##/utils/utils.js';
 import { AzureDevopsApi } from '##/use-cases/version-control/azure-devops/apis/azure-devops.api.js';
 
 import { AZURE_PULL_REQUESTS_RESPONSE, SERVER_PULL_REQUESTS_RESPONSE } from './pull-requests.mock.js';
-import { SERVER_ERROR_MESSAGE, STATUS_CODE } from '##/constants/index.js';
+import { SERVER_ERROR_MESSAGE, STATUS_CODE } from '##/constants/constants.js';
 import { runDatePaginationValidationTests } from '../../common-tests/date-pagination-tests.js';
 
 const { invalidRepositoryDetails, invalidAzureToken, dataNotFound } = AzureDevopsApi;
